requirements.txt
flask
mysql-connector-python
waitress
paste
yaml
pandas
